drivers: arm: stm32: adc Add Shared IRQ
Certain stm32 soc's have a single shared irq for all adc channels
on those soc's only a single channel is supported.

Added a Kconfig that enables a shared irq for stm32f2x, stm32f4x,
stm32f7x soc's. The shared IRQ uses a flag to limit the number of
interrupts defined to only 1. A shared irq handler is added which
determines which ADC instance the interrupt is for, it then calls
into the existing interrupt.

A non-required dt config was also added to enable configuration of
the prescalar for these soc's in the device tree, currently it is
hardcoded in the driver.

Signed-off-by: Marius Scholtz <mariuss@ricelectronics.com>